Color picker working.

+++ EclipseEngine/Shaders/MeshPicker3D.frag
@@ -25,7 +25,7 @@ layout(push_constant) uniform SceneData
     mat4 proj;
     mat4 view;
     vec3 CameraPos;
-    ivec3 MeshColorID;
+    vec3 MeshColorID;
     float Timer;
 } sceneData;
